refactor method/group/* into one file, rename to group/po-box
