--- spec/group/add-pobox.js
@@ -1,9 +1,0 @@
-const { poBox } = require('../lib/field-types')
-
-module.exports = {
-  type: 'group/poBox',
-  tangle: 'poBox',
-  props: {
-    keys: poBox
-  }
-}
