+++ spec/group/po-box.js
@@ -1,0 +1,9 @@
+const { poBoxInfo } = require('../lib/field-types')
+
+module.exports = {
+  type: 'group/po-box',
+  tangle: 'poBox',
+  props: {
+    keys: poBoxInfo
+  }
+}
